pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bibslim"
version = "0.1.0"
description = "BibTeX slimmer: abbreviates venues, shortens authors, normalizes titles/pages/DOIs, and prunes noisy fields."
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "François", email = "francois@example.com" }]
license = { file = "LICENSE" }
dependencies = [
  "bibtexparser>=1.4.0",
  "python-slugify>=8.0.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.1.13",
  "mypy>=1.6",
]

[project.scripts]
bibslim = "bibslim.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"bibslim" = ["data/*.yml"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.9"
warn_unused_ignores = false
strict = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
testpaths = ["tests"]
